Improve copy-from support for bionics and expand checks

+++ src/bionics_ui.cpp
@@ -392,10 +392,11 @@ static void draw_connectors( const catacurses::window &win, const point &start,
     const int LIST_START_Y = 7;
     // first: pos_y, second: occupied slots
     std::vector<std::pair<int, size_t>> pos_and_num;
-    for( const std::pair<const bodypart_str_id, size_t> &elem : bio_id->occupied_bodyparts ) {
+    for( const std::pair<const bodypart_str_id, int> &elem : bio_id->occupied_bodyparts ) {
         auto pos = bp_to_pos.find( elem.first );
         if( pos != bp_to_pos.end() ) {
-            pos_and_num.emplace_back( static_cast<int>( pos->second ) + LIST_START_Y, elem.second );
+            pos_and_num.emplace_back( static_cast<int>( pos->second ) + LIST_START_Y,
+                                      static_cast<size_t>( elem.second ) );
         }
     }
     if( pos_and_num.empty() || !get_option < bool >( "CBM_SLOTS_ENABLED" ) ) {
@@ -686,7 +687,7 @@ void player::power_bionics()
                                      pos_x - 2, bio_id, bp_to_pos );
 
                     // redraw highlighted (occupied) body parts
-                    for( const std::pair<const bodypart_str_id, size_t> &elem : bio_id->occupied_bodyparts ) {
+                    for( const std::pair<const bodypart_str_id, int> &elem : bio_id->occupied_bodyparts ) {
                         const int i = bp_to_pos[elem.first];
                         mvwprintz( wBio, point( pos_x, i + list_start_y ), c_yellow, bps[i] );
                     }
